% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare_levy.R
\name{check_upper_bound}
\alias{check_upper_bound}
\title{Check whether an income should be tapered}
\usage{
check_upper_bound(income, lower_bound, rate = 0.02, taper_rate = 0.1)
}
\arguments{
\item{income}{Income}

\item{lower_bound}{Level of the lower bound}

\item{rate}{Medicare Levy Rate}

\item{taper_rate}{Low income taper rate}
}
\value{
Logical. Whether the income is below the maximum taper threshold
}
\description{
Check whether an income should be tapered
}
