% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecs.R
\name{hecs}
\alias{hecs}
\title{HECS repayment calculator}
\usage{
hecs(income, debt, brackets = c(0, 51957, 57730, 64307, 70882, 74608, 80198,
  86856, 91426, 100614, 107214), rates = c(0, 0.02, 0.04, 0.045, 0.05, 0.055,
  0.06, 0.065, 0.07, 0.075, 0.08))
}
\arguments{
\item{income}{HECS Repayment Income}

\item{debt}{Outstanding balance}

\item{brackets}{}

\item{rates}{}
}
\value{
The amount of HECS repayments in a financial year
}
\description{
\code{hecs} calculates the amount of HECS to be paid in a financial year
}
\details{
HECS repayment income is different to your taxable income. It is calculated as:
\itemize{
  \item{your taxable income for an income year, plus}
  \item{your total net investment losses, plus}
  \item{any total reportable fringe benefit amounts shown on your PAYG payment summary; plus}
  \item{reportable super contributions; and}
  \item{any exempt foreign employment income from the current income year.}
}
}
\examples{

}
