% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_tax.R
\name{income_tax}
\alias{income_tax}
\title{Calculate income or superannuation tax liability. Defaults to 2012-13 tax
scales.}
\usage{
income_tax(income, rates = c(0, 0.19, 0.325, 0.37, 0.45),
  brackets = c(18200, 37000, 80000, 180000, Inf))
}
\arguments{
\item{income}{Numeric. Gross income.}

\item{rates}{Numeric vector. Rate for each bracket.}

\item{brackets}{Numeric vector. Top of each tax bracket.}
}
\description{
Calculate income or superannuation tax liability. Defaults to 2012-13 tax
scales.
}
\examples{
# Calculate tax liability for taxpayer with gross income of 50,000
income_tax(50000)
}
