% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmito.R
\name{lmito}
\alias{lmito}
\title{Low and Middle Income Tax Offset}
\usage{
lmito(income, initial_value = 200, increased_value = 530,
  taper_one_start = 37000, taper_two_start = 90000, taper_in_rate = 0.03,
  taper_out_rate = 0.015)
}
\arguments{
\item{income}{Numeric. Gross income.}

\item{initial_value}{Numeric. The value of the first step of the offset.}

\item{increased_value}{Numeric. THe value of the second step of the offset.}

\item{taper_one_start}{Numeric. The point at which the offset tapers up.}

\item{taper_two_start}{Numeric. The point at which the offset tapers down.}

\item{taper_in_rate}{Numeric. The rate at which the offset tapers up.}

\item{taper_out_rate}{Numeric. The rate at which the offset tapers out.}
}
\description{
Low and Middle Income Tax Offset
}
\examples{
lmito(75000)
}
