% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_tax_function.R
\name{make_tax_function}
\alias{make_tax_function}
\title{Make a function to calculate marginal income tax rates}
\usage{
make_tax_function(rates, brackets, super = FALSE)
}
\arguments{
\item{rates}{Numeric vector. Personal income tax rates.}

\item{brackets}{Numeric vector. Personal income tax brackets.}

\item{super}{Logical. Whether this is a superannuation tax calculator.}
}
\value{
A function to calculate tax rates.
}
\description{
This closure returns functions that can be easily customised to test
different marginal income tax rates.
}
\examples{
# Make a tax calculator using the 2012-13 income tax rates and brackets
pit_12_13 <- make_tax_function(rates = c(0, .19, .325, .37, .45),
                               brackets = c(18200,37000,80000,180000,Inf))
}

