% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/medicare_levy.R
\name{medicare_levy}
\alias{medicare_levy}
\title{Calculate the medicare levy owed.}
\usage{
medicare_levy(income, levy_rate = 0.02, shading_brackets = c(20896, 26121),
  shading_rates = c(0, 0.1))
}
\arguments{
\item{income}{Numeric. Gross income.}

\item{levy_rate}{Numeric. Rate at which the medicare levy is paid.}

\item{shading_brackets}{Numeric vector. Income level at which the
medicare levy is 'shaded' or phased in.}

\item{shading_rates}{Numeric vector. The rates at which the medicare levy
  is 'shaded' or phased in.}
}
\value{
Numeric, the amount of medicare levy owed
}
\description{
The medicare levy is not a marginal tax, it's a flat two per cent tax
with exemptions for low income earners. If you're below the lowest
'shading bracket' then you're exempt from the medicare levy, between the
lowest shading bracket and the top shading bracket the levy is a 10 %
marginal tax of income over the bottom shading bracket. Once you get over
the top shading bracket it becomes a flat 2 per cent tax.
}
\examples{
medicare_levy(20000)
}

