% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare_levy_progressive.R
\name{ml_progressive}
\alias{ml_progressive}
\title{Progessive medicare levy}
\usage{
ml_progressive(x, cutoff = 87000, base_rate = 0.02, increase_rate = 0.025,
  ...)
}
\arguments{
\item{x}{Income}

\item{cutoff}{Where does the progressive step kick in}

\item{base_rate}{Normal rate}

\item{increase_rate}{Extra rate}

\item{...}{Additional parameters to medicare levy function}
}
\value{
Progressive medicare levy owed in dollars
}
\description{
Progessive medicare levy
}
\examples{
ml_progressive(50000)
}
