% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lito.R
\name{lito}
\alias{lito}
\title{Calculate the low income tax offset. This is deducted from non-refundable
tax offset that reduces individuals' tax owed.}
\usage{
lito(income, amount = 445, taper_brackets = c(37000, 66667),
  taper_rate = 0.015)
}
\arguments{
\item{income}{Numeric. Gross income.}

\item{amount}{Numeric. The untapered low income tax offset amount.}

\item{taper_brackets}{Numeric vector.}
}
\description{
Calculate the low income tax offset. This is deducted from non-refundable
tax offset that reduces individuals' tax owed.
}
\examples{
lito(18000)
}
