% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare_levy.R
\name{medicare_levy}
\alias{medicare_levy}
\title{Calculate the amount of medicare levy owed}
\usage{
medicare_levy(income, rate = 0.02, taper_rate = 0.1,
  lower_bound_single = 21335, lower_bound_couple = 36001,
  spouse_income = 0, n_children = 0, children_reduction = 3306,
  rel_status = "single")
}
\arguments{
\item{income}{Income of the person for whom the levy is being calulated}

\item{rate}{Medicare levy rate}

\item{taper_rate}{How fast should the medicare levy low income reduction taper out}

\item{lower_bound_single}{The start of the taper out of the low income reduction (single)}

\item{lower_bound_couple}{The start of the taper out of the low income reduction (family)}

\item{spouse_income}{Income of the person's spouse}

\item{n_children}{Number of children}

\item{children_reduction}{How much should the lower_bound_couple increase by for each child}

\item{rel_status}{Relationship status, must either be 'single' or 'couple'}
}
\value{
Amount of medicare levy owed
}
\description{
Calculate the amount of medicare levy owed
}
\examples{
medicare_levy(100000)
medicare_levy(44000, spouse_income = 0, rel_status = 'couple', lower_bound_single = 20896,
              lower_bound,couple = 35261, children_reduction = 3238)
}
